move AppLayout to components/layouts
