+++ cyclops-ui/src/App.tsx
@@ -4,7 +4,7 @@ import {
 } from "react-router-dom"
 import routes from "./routes"
 import Page404 from "./components/pages/Page404"
-import AppLayout from "./AppLayout";
+import AppLayout from "./components/layouts/AppLayout";
 
 export default function App() {
   const router = createBrowserRouter([
